Namespace all selects
